# Conflicts:
#	WHATSNEW
#	_scripts/travis/travis-stage-fingerprints.sh
#	_scripts/update-changelogs
#	examples/inet/nclients/basicHTTP.ini
#	examples/inet/nclients/nclients2.ini
#	examples/inet/nclients/omnetpp-ping.ini
#	examples/inet/nclients/omnetpp.ini
#	examples/inet/shutdownrestart/omnetpp.ini
#	examples/ipv6/demonetworketh/omnetpp.ini
#	examples/ipv6/ipv6nclients/omnetpp.ini
#	examples/ipv6/nclients/omnetpp.ini
#	examples/mobileipv6/omnetpp.ini
#	examples/sctp/nclients/omnetpp.ini
#	src/inet/applications/httptools/server/HttpServerEvilA.ned
#	src/inet/applications/pingapp/PingApp.ned
#	src/inet/common/INETDefs.h
#	src/inet/linklayer/csma/ChangeLog
#	src/inet/linklayer/ieee80211/mac/channelaccess/Dcaf.cc
#	src/inet/networklayer/multi/ChangeLog
#	src/inet/node/httptools/DirectHost.ned
#	src/inet/physicallayer/unitdisk/UnitDiskTransmitter.ned
#	tests/features/featuretest
#	tests/module/diffserv_dscpmarker_1.test
#	tests/module/lo0_IPv4.test
#	tests/module/lo0_IPv6.test
#	tests/module/ospf_1_area.test
#	tests/module/ospf_1_area_HostInterface.test
#	tests/module/ospf_1_area_lifecycle.test
#	tests/module/ospf_backbone_and_2_areas.test
#	tests/module/ospf_backbone_and_2_areas_HostInterface.test
#	tests/module/ospf_backbone_and_2_stub.test
#	tests/module/ospf_backbone_and_3_areas_with_virtual_link.test
#	tests/module/rip_1.test
#	tests/module/runtest
#	tests/module/sctp_streamReset.test
#	tests/module/sctp_streams.test
#	tests/module/tcp_algorithm_dumb.test
#	tests/module/tcp_algorithm_newreno.test
#	tests/module/tcp_algorithm_reno.test
#	tests/module/tcp_algorithm_tahoe.test
#	tests/module/tcp_algorithm_vegas.test
#	tests/module/tcp_algorithm_westwood.test

+++ src/inet/transportlayer/sctp/SCTPAssociationRcvMessage.cc
@@ -845,7 +845,7 @@ bool SCTPAssociation::processCookieEchoArrived(SCTPCookieEchoChunk *cookieEcho, 
 {
     bool trans = false;
     SCTPCookie *cookie = check_and_cast<SCTPCookie *>(cookieEcho->getStateCookie());
-    if (cookie->getCreationTime() + (int32)sctpMain->par("validCookieLifetime") < simTime()) {
+    if (cookie->getCreationTime() + sctpMain->par("validCookieLifetime") < simTime()) {
         EV_INFO << "stale Cookie: sendAbort\n";
         sendAbort();
         delete cookie;
